+++ extensions/sourcegraph/src/components/CloudCommand.tsx
@@ -1,9 +1,7 @@
-import { ApolloProvider } from "@apollo/client";
 import { useEffect } from "react";
 
 import checkAuthEffect from "../hooks/checkAuthEffect";
 import { Sourcegraph, sourcegraphCloud } from "../sourcegraph";
-import { newApolloClient } from "../sourcegraph/gql/apollo";
 
 /**
  * SelfHostedCommand wraps the given command with the configuration for Sourcegraph Cloud.
@@ -13,9 +11,5 @@ export default function CloudCommand({ Command }: { Command: React.FunctionCompo
 
   useEffect(checkAuthEffect(src));
 
-  return (
-    <ApolloProvider client={newApolloClient(src)}>
-      <Command src={src} />
-    </ApolloProvider>
-  );
+  return <Command src={src} />;
 }