+++ extensions/sourcegraph/src/components/SelfHostedCommand.tsx
@@ -1,10 +1,9 @@
-import { ApolloProvider } from "@apollo/client";
-import { ActionPanel, Detail, Action } from "@raycast/api";
+import { ActionPanel, Detail, Action, Icon } from "@raycast/api";
 import { useEffect } from "react";
 
 import checkAuthEffect from "../hooks/checkAuthEffect";
+import { bold } from "../markdown";
 import { sourcegraphSelfHosted, Sourcegraph } from "../sourcegraph";
-import { newApolloClient } from "../sourcegraph/gql/apollo";
 
 /**
  * SelfHostedCommand wraps the given command with the configuration for a self-hosted
@@ -17,6 +16,7 @@ export default function SelfHostedCommand({ Command }: { Command: React.Function
     <ActionPanel>
       <Action.OpenInBrowser
         title="Open Setup Guide"
+        icon={Icon.Globe}
         url="https://github.com/raycast/extensions/tree/main/extensions/sourcegraph#setup"
       />
     </ActionPanel>
@@ -27,7 +27,9 @@ export default function SelfHostedCommand({ Command }: { Command: React.Function
     return (
       <Detail
         navigationTitle="No Sourcegraph Self-Hosted instance configured"
-        markdown={`**⚠️ No Sourcegraph Sourcegraph Self-Hosted instance configured** - set one up in the extension preferences!\n\n${tryCloudMessage}`}
+        markdown={`${bold(
+          `⚠️ No Sourcegraph Sourcegraph Self-Hosted instance configured`
+        )} - set one up in the extension preferences!\n\n${tryCloudMessage}`}
         actions={helpActions}
       />
     );
@@ -38,7 +40,9 @@ export default function SelfHostedCommand({ Command }: { Command: React.Function
     return (
       <Detail
         navigationTitle="Invalid Sourcegraph Self-Hosted URL"
-        markdown={`**⚠️ Sourcegraph Self-Hosted URL '${src.instance}' is invalid:** ${e}\n\nUpdate it in the extension preferences!\n\n${tryCloudMessage}`}
+        markdown={`${bold(
+          `⚠️ Sourcegraph Self-Hosted URL '${src.instance}' is invalid:`
+        )} ${e}\n\nUpdate it in the extension preferences!\n\n${tryCloudMessage}`}
         actions={helpActions}
       />
     );
@@ -47,7 +51,9 @@ export default function SelfHostedCommand({ Command }: { Command: React.Function
     return (
       <Detail
         navigationTitle="Invalid Sourcegraph Self-Hosted access token"
-        markdown={`**⚠️ A token is required for Sourcegraph Self-Hosted instance '${src.instance}'** - add an access token for Sourcegraph Self-Hosted in the extension preferences!\n\n${tryCloudMessage}`}
+        markdown={`${bold(
+          `⚠️ A token is required for Sourcegraph Self-Hosted instance '${src.instance}'`
+        )} - add an access token for Sourcegraph Self-Hosted in the extension preferences!\n\n${tryCloudMessage}`}
         actions={helpActions}
       />
     );
@@ -55,9 +61,5 @@ export default function SelfHostedCommand({ Command }: { Command: React.Function
 
   useEffect(checkAuthEffect(src));
 
-  return (
-    <ApolloProvider client={newApolloClient(src)}>
-      <Command src={src} />
-    </ApolloProvider>
-  );
+  return <Command src={src} />;
 }