+++ extensions/sourcegraph/src/sourcegraph/gql/queries.ts
@@ -80,6 +80,8 @@ const BATCH_CHANGE_FIELDS = gql`
       open
       closed
       failed
+      unpublished
+      draft
     }
   }
 `;
@@ -130,6 +132,26 @@ export const GET_CHANGESETS = gql`
       changesets {
         nodes {
           ...ChangesetFields
+        }
+      }
+    }
+  }
+`;
+
+export const BLOB_CONTENTS = gql`
+  fragment BlobContents on GitBlob {
+    content
+    binary
+  }
+`;
+
+export const GET_FILE_CONTENTS = gql`
+  ${BLOB_CONTENTS}
+  query GetFileContents($repo: String!, $rev: String!, $path: String!) {
+    repository(name: $repo) {
+      commit(rev: $rev) {
+        blob(path: $path) {
+          ...BlobContents
         }
       }
     }