+++ extensions/sourcegraph/CHANGELOG.md
@@ -10,6 +10,7 @@
 - **search**: The default context is now explicitly pre-populated in the search bar, instead of being implicit.
 - **search**: Potential fix for issue with connecting to self-hosted Sourcegraph instances. ([raycast/extensions#1126](https://github.com/raycast/extensions/issues/1126))
 - **search**: You can now apply suggestions directly with an action.
+- **search**: Path result previews (e.g. from `type:path` or `select:path`) now render a preview for certain file types (currently only Markdown and Golang).
 - **prefs**: Fixed issue with self-hosted instance URLs that have trailing slashes. ([#12](https://github.com/bobheadxi/raycast-sourcegraph/pull/12))
 
 ## [Bug fixes](https://github.com/raycast/extensions/pull/919) - 2022-02-15